Remove arrow2
