+++ Cargo.toml
@@ -15,14 +15,7 @@ rust-version = "1.62"
 crate-type = ["cdylib", "rlib"]
 
 [features]
-default = ["arrow1", "all_compressions", "reader", "writer"]
-arrow1 = ["dep:arrow", "dep:parquet", "dep:bytes", "arrow-wasm/arrow1"]
-arrow2 = [
-  "dep:arrow2",
-  "dep:parquet2",
-  "dep:serde-wasm-bindgen",
-  "arrow-wasm/arrow2",
-]
+default = ["all_compressions", "reader", "writer", "async"]
 reader = []
 writer = []
 async = [
@@ -33,23 +26,21 @@ async = [
   "dep:wasm-streams",
   "dep:async-compat",
   "dep:async-stream",
-  "parquet?/async",
+  "parquet/async",
 ]
 debug = ["console_error_panic_hook"]
 
-brotli = ["parquet?/brotli", "parquet2?/brotli"]
-gzip = ["parquet?/flate2", "parquet2?/gzip"]
-snappy = ["parquet?/snap", "parquet2?/snappy"]
-zstd = ["parquet?/zstd", "parquet2?/zstd", "dep:zstd"]
+brotli = ["parquet/brotli"]
+gzip = ["parquet/flate2"]
+snappy = ["parquet/snap"]
+zstd = ["parquet/zstd", "dep:zstd"]
 # LZ4 not yet supported on parquet crate
-lz4 = ["parquet2?/lz4_flex"]
+# lz4 = ["parquet2?/lz4_flex"]
 
-all_compressions = ["brotli", "gzip", "snappy", "zstd", "lz4"]
+all_compressions = ["brotli", "gzip", "snappy", "zstd"] # "lz4"
 
 # Full list of available features
 full = [
-  "arrow1",
-  "arrow2",
   "async",
   "debug",
   "all_compressions",
@@ -76,26 +67,19 @@ js-sys = "0.3.60"
 getrandom = { version = "0.2.6", features = ["js"] }
 thiserror = "1.0"
 
-arrow-wasm = { git = "https://github.com/kylebarron/arrow-wasm", rev = "38aa646f6bf0ec7fe3eacc0d9e47671ac82de7b5" }
+# c7a0db3b2f7034878796939ad46955450cb2c2e7
+arrow-wasm = { git = "https://github.com/kylebarron/arrow-wasm", rev = "c7a0db3b2f7034878796939ad46955450cb2c2e7" }
 # arrow-wasm = { path = "/Users/kyle/github/rust/arrow-wasm" }
 
-arrow2 = { version = "0.18", optional = true, features = [
-  "io_ipc",
-  "io_parquet",
-] }
-parquet2 = { version = "0.17", default_features = false, optional = true }
-
-arrow = { version = "50", default-features = false, optional = true, features = [
+arrow = { version = "50", default-features = false, features = [
   "ipc",
   "ffi",
 ] }
-parquet = { version = "50", default-features = false, optional = true, features = [
+parquet = { version = "50", default-features = false, features = [
   "arrow",
   "base64",
 ] }
-bytes = { version = "1", optional = true }
-
-serde-wasm-bindgen = { version = "0.6", optional = true }
+bytes = "1"
 
 wasm-bindgen-futures = { version = "0.4.30", optional = true }
 futures = { version = "0.3", optional = true }
@@ -124,17 +108,13 @@ features = [
   'Window',
   "Document",
   "Element",
-  "File"
+  "File",
 ]
 
 [dev-dependencies]
 wasm-bindgen-test = "0.3.31"
 
 [package.metadata.cargo-all-features]
-
-# In practice, no one would ever use these pairs together, so we don't test their
-# intersections
-skip_feature_sets = [["arrow1", "arrow2"], ["arrow1", "lz4"]]
 
 # If your crate has a large number of optional dependencies, skip them for speed
 skip_optional_dependencies = true
