--- src/arrow1/mod.rs
@@ -1,11 +1,0 @@
-pub mod error;
-pub mod metadata;
-#[cfg(feature = "reader")]
-pub mod reader;
-#[cfg(all(feature = "reader", feature = "async"))]
-pub mod reader_async;
-pub mod wasm;
-#[cfg(feature = "writer")]
-pub mod writer;
-#[cfg(feature = "writer")]
-pub mod writer_properties;
