+++ templates/package.json
@@ -12,61 +12,28 @@
   "files": [
     "*"
   ],
-  "module": "bundler/arrow2.js",
-  "types": "bundler/arrow2.d.ts",
-  "sideEffects": [
-    "./arrow1.js",
-    "./snippets/*"
-  ],
+  "module": "bundler/parquet_wasm.js",
+  "types": "bundler/parquet_wasm.d.ts",
+  "sideEffects": [],
   "keywords": [
     "parquet",
     "webassembly",
     "arrow"
   ],
   "exports": {
-    "./arrow1": {
-      "import": {
-        "types": "./esm/arrow1.d.ts",
-        "import": "./esm/arrow1.js"
-      },
-      "browser": {
-        "types": "./esm/arrow1.d.ts",
-        "import": "./esm/arrow1.js"
-      },
-      "node": {
-        "types": "./node/arrow1.d.ts",
-        "require": "./node/arrow1.js",
-        "import": "./node/arrow1.js"
-      }
-    },
-    "./arrow2": {
-      "import": {
-        "types": "./esm/arrow2.d.ts",
-        "import": "./esm/arrow2.js"
-      },
-      "browser": {
-        "types": "./esm/arrow2.d.ts",
-        "import": "./esm/arrow2.js"
-      },
-      "node": {
-        "types": "./node/arrow2.d.ts",
-        "require": "./node/arrow2.js",
-        "import": "./node/arrow2.js"
-      }
-    },
     ".": {
       "import": {
-        "types": "./esm/arrow1.d.ts",
-        "import": "./esm/arrow1.js"
+        "types": "./esm/parquet_wasm.d.ts",
+        "import": "./esm/parquet_wasm.js"
       },
       "browser": {
-        "types": "./esm/arrow1.d.ts",
-        "import": "./esm/arrow1.js"
+        "types": "./esm/parquet_wasm.d.ts",
+        "import": "./esm/parquet_wasm.js"
       },
       "node": {
-        "types": "./node/arrow1.d.ts",
-        "require": "./node/arrow1.js",
-        "import": "./node/arrow1.js"
+        "types": "./node/parquet_wasm.d.ts",
+        "require": "./node/parquet_wasm.js",
+        "import": "./node/parquet_wasm.js"
       }
     }
   }
